+++ src/Modules/Guilds/Hooks/useProposalCalls.ts
@@ -1,14 +1,14 @@
 import { useState, useEffect, useMemo } from 'react';
 import { useTheme } from 'styled-components';
 import { useTranslation } from 'react-i18next';
-import { bulkDecodeCallsFromOptions } from '../contracts/useDecodedCall';
+import { bulkDecodeCallsFromOptions } from 'hooks/Guilds/contracts/useDecodedCall';
 import { decodeCall } from 'hooks/Guilds/contracts/useDecodedCall';
 import { useProposal } from './useProposal';
-import { useVotingResults } from 'hooks/Guilds/guild/useVotingResults';
+import { useVotingResults } from 'Modules/Guilds/Hooks/useVotingResults';
 import { Call, Option } from 'components/ActionsBuilder/types';
 import { ZERO_HASH } from 'utils';
-import useProposalMetadata from '../useProposalMetadata';
-import { useRichContractRegistry } from '../contracts/useRichContractRegistry';
+import useProposalMetadata from 'hooks/Guilds/useProposalMetadata';
+import { useRichContractRegistry } from 'hooks/Guilds/contracts/useRichContractRegistry';
 import { ERC20_APPROVE_SIGNATURE } from 'utils';
 import { useNetwork } from 'wagmi';
 import { getBigNumberPercentage } from 'utils/bnPercentage';
