+++ src/Modules/Guilds/Wrappers/GuildSidebarWrapper.tsx
@@ -1,20 +1,20 @@
 import { useTypedParams } from 'Modules/Guilds/Hooks/useTypedParams';
-import { useGuildConfig } from 'hooks/Guilds/guild/useGuildConfig';
-import useGuildMemberTotal from 'hooks/Guilds/guild/useGuildMemberTotal';
-import { useVotingPowerOf } from 'hooks/Guilds/guild/useVotingPowerOf';
+import { useGuildConfig } from 'Modules/Guilds/Hooks/useGuildConfig';
+import useGuildMemberTotal from 'Modules/Guilds/Hooks/useGuildMemberTotal';
+import { useVotingPowerOf } from 'Modules/Guilds/Hooks/useVotingPowerOf';
 import { GuildSidebar } from 'components/GuildSidebar';
 import { MemberActions } from 'components/GuildSidebar/MemberActions';
 import { GuestActions } from 'components/GuildSidebar/GuestActions';
 import { useERC20Info } from 'hooks/Guilds/erc20/useERC20Info';
-import { useVoterLockTimestamp } from 'hooks/Guilds/guild/useVoterLockTimestamp';
-import useGuildImplementationType from 'hooks/Guilds/guild/useGuildImplementationType';
+import { useVoterLockTimestamp } from 'Modules/Guilds/Hooks/useVoterLockTimestamp';
+import useGuildImplementationType from 'Modules/Guilds/Hooks/useGuildImplementationType';
 import { useTransactions } from 'contexts/Guilds';
 import { useERC20Guild } from 'hooks/Guilds/contracts/useContract';
 import { formatUnits } from 'ethers/lib/utils';
-import useVotingPowerPercent from 'hooks/Guilds/guild/useVotingPowerPercent';
+import useVotingPowerPercent from 'Modules/Guilds/Hooks/useVotingPowerPercent';
 import { useState } from 'react';
 import { WalletModal } from 'components/Web3Modals';
-import useTotalLocked from 'hooks/Guilds/guild/useTotalLocked';
+import useTotalLocked from 'Modules/Guilds/Hooks/useTotalLocked';
 import StakeTokensModalWrapper from './StakeTokensModalWrapper';
 import { useAccount, useEnsAvatar, useEnsName } from 'wagmi';
 import { isReadOnly } from 'provider/wallets';
