+++ src/Modules/Guilds/Wrappers/ProposalVoteCardWrapper.tsx
@@ -1,10 +1,10 @@
 import { useERC20Guild } from 'hooks/Guilds/contracts/useContract';
-import { useProposal } from 'hooks/Guilds/guild/useProposal';
+import { useProposal } from 'Modules/Guilds/Hooks/useProposal';
 import useProposalMetadata from 'hooks/Guilds/useProposalMetadata';
-import useSnapshotId from 'hooks/Guilds/guild/useSnapshotId';
-import { useVotingPowerOf } from 'hooks/Guilds/guild/useVotingPowerOf';
-import { useVotingResults } from 'hooks/Guilds/guild/useVotingResults';
-import useVotingPowerPercent from 'hooks/Guilds/guild/useVotingPowerPercent';
+import useSnapshotId from 'Modules/Guilds/Hooks/useSnapshotId';
+import { useVotingPowerOf } from 'Modules/Guilds/Hooks/useVotingPowerOf';
+import { useVotingResults } from 'Modules/Guilds/Hooks/useVotingResults';
+import useVotingPowerPercent from 'Modules/Guilds/Hooks/useVotingPowerPercent';
 import useTimedRerender from 'hooks/Guilds/time/useTimedRerender';
 import { useTypedParams } from 'Modules/Guilds/Hooks/useTypedParams';
 import { ProposalVoteCard } from 'components/ProposalVoteCard';
