+++ src/Modules/Guilds/pages/Proposal/Proposal.tsx
@@ -1,4 +1,4 @@
-import { useProposal } from 'hooks/Guilds/guild/useProposal';
+import { useProposal } from 'Modules/Guilds/Hooks/useProposal';
 import AddressButton from 'components/AddressButton/AddressButton';
 import { ProposalDescription } from 'components/ProposalDescription';
 import { ProposalInfoCard } from 'components/ProposalInfoCard';
@@ -7,10 +7,10 @@ import { IconButton } from 'components/primitives/Button';
 import { UnstyledLink } from 'components/primitives/Links';
 import { useTypedParams } from 'Modules/Guilds/Hooks/useTypedParams';
 import { GuildAvailabilityContext } from 'contexts/Guilds/guildAvailability';
-import { useGuildProposalIds } from 'hooks/Guilds/guild/useGuildProposalIds';
-import useTotalLocked from 'hooks/Guilds/guild/useTotalLocked';
-import useSnapshotId from 'hooks/Guilds/guild/useSnapshotId';
-import useProposalCalls from 'hooks/Guilds/guild/useProposalCalls';
+import { useGuildProposalIds } from 'Modules/Guilds/Hooks/useGuildProposalIds';
+import useTotalLocked from 'Modules/Guilds/Hooks/useTotalLocked';
+import useSnapshotId from 'Modules/Guilds/Hooks/useSnapshotId';
+import useProposalCalls from 'Modules/Guilds/Hooks/useProposalCalls';
 import { Loading } from 'components/primitives/Loading';
 import { Result, ResultState } from 'components/Result';
 import React, { useContext } from 'react';
@@ -20,10 +20,10 @@ import ProposalVoteCardWrapper from 'Modules/Guilds/Wrappers/ProposalVoteCardWra
 import { ExecuteButton } from 'components/ExecuteButton';
 import { useProposalState } from 'hooks/Guilds/useProposalState';
 import useExecutable from 'hooks/Guilds/useExecutable';
-import { useGuildConfig } from 'hooks/Guilds/guild/useGuildConfig';
+import { useGuildConfig } from 'Modules/Guilds/Hooks/useGuildConfig';
 import { ProposalState } from 'types/types.guilds.d';
 import useProposalMetadata from 'hooks/Guilds/useProposalMetadata';
-import useVotingPowerPercent from 'hooks/Guilds/guild/useVotingPowerPercent';
+import useVotingPowerPercent from 'Modules/Guilds/Hooks/useVotingPowerPercent';
 import { ActionsBuilder } from 'components/ActionsBuilder';
 import { useAccount } from 'wagmi';
 import { isReadOnly } from 'provider/wallets';
