+++ src/components/ActionsBuilder/SupportedActions/RepMint/RepMintEditor.tsx
@@ -8,8 +8,8 @@ import useENSAvatar from 'hooks/Guilds/ens/useENSAvatar';
 import { MAINNET_ID } from 'utils';
 import { ReactComponent as Info } from 'assets/images/info.svg';
 import useBigNumberToNumber from 'hooks/Guilds/conversions/useBigNumberToNumber';
-import { useTotalSupply } from 'hooks/Guilds/guild/useTotalSupply';
-import { useTokenData } from 'hooks/Guilds/guild/useTokenData';
+import { useTotalSupply } from 'Modules/Guilds/Hooks/useTotalSupply';
+import { useTokenData } from 'Modules/Guilds/Hooks/useTokenData';
 import { Tooltip } from 'components/Tooltip';
 import { useTranslation } from 'react-i18next';
 import { ethers } from 'ethers';
