+++ src/components/StakeTokensModal/types.ts
@@ -1,7 +1,7 @@
 import { BigNumber, providers } from 'ethers';
 import { ERC20Info } from 'hooks/Guilds/erc20/useERC20Info';
 import { ERC20, ERC20Guild } from 'types/contracts';
-import { GuildConfigProps } from 'hooks/Guilds/guild/useGuildConfig';
+import { GuildConfigProps } from 'Modules/Guilds/Hooks/useGuildConfig';
 import React from 'react';
 export interface StakeTokensModalProps {
   isOpen: boolean;
