+++ src/hooks/Guilds/erc20/useERC20Info.ts
@@ -40,8 +40,10 @@ export const useERC20Info = (contractAddress: string): ERC20InfoHook => {
       },
     ],
   });
-  if (!data || data.every(v => v === null))
+  if (!data || data.every(v => v === null)) {
+    console.log('useERC20Info', data);
     return { data: undefined, isError, isLoading };
+  }
   return {
     data: {
       name: data[0].toString(),
